rename manager to cnb
