+++ lib/modules/manager/api.ts
@@ -12,7 +12,6 @@ import * as bicep from './bicep';
 import * as bitbucketPipelines from './bitbucket-pipelines';
 import * as bitrise from './bitrise';
 import * as buildkite from './buildkite';
-import * as buildpacks from './buildpacks';
 import * as bun from './bun';
 import * as bunVersion from './bun-version';
 import * as bundler from './bundler';
@@ -21,6 +20,7 @@ import * as cargo from './cargo';
 import * as cdnurl from './cdnurl';
 import * as circleci from './circleci';
 import * as cloudbuild from './cloudbuild';
+import * as cnb from './cnb';
 import * as cocoapods from './cocoapods';
 import * as composer from './composer';
 import * as conan from './conan';
@@ -125,7 +125,7 @@ api.set('cargo', cargo);
 api.set('cdnurl', cdnurl);
 api.set('circleci', circleci);
 api.set('cloudbuild', cloudbuild);
-api.set('buildpacks', buildpacks);
+api.set('cnb', cnb);
 api.set('cocoapods', cocoapods);
 api.set('composer', composer);
 api.set('conan', conan);
