+++ lib/modules/manager/cnb/__fixtures__/empty_project.toml
