+++ lib/modules/manager/cnb/__fixtures__/project.toml
