+++ lib/modules/manager/cnb/extract.ts
