+++ lib/modules/manager/cnb/index.ts
