+++ lib/modules/manager/cnb/schema.ts
